refactoring of code

+++ data/base_dataset.py
@@ -5,7 +5,7 @@ It also includes common transformation functions (e.g., get_transform, __scale_w
 import random
 import numpy as np
 import torch.utils.data as data
-from torch import float16,float32,uint8
+from torch import float16,float32,uint8,float64
 from PIL import Image
 import torchvision.transforms.v2 as transforms
 from abc import ABC, abstractmethod
@@ -106,7 +106,7 @@ def get_transform(opt, params=None, grayscale=False, method=transforms.Interpola
 
     if convert:
         #transform_list += [transforms.ToTensor()]
-        transform_list.extend([transforms.ToImage(), transforms.ToDtype(float32, scale=True)])
+        transform_list.extend([transforms.ToImage(), transforms.ToDtype(float16, scale=True)])
         if grayscale:
             transform_list += [transforms.Normalize((0.5,), (0.5,))]
         else:
